cicd/imagery/requirements.txt
requests==2.26.0
pyproj==3.3.0
Pillow==8.4.0